fix: errors related with moving streamApiLow
