+++ jni-wrappers/privmx-endpoint/src/cpp/modules/StreamApiLow.cpp
@@ -26,77 +26,33 @@ StreamApiLow *getStreamApi(JniContextUtils &ctx, jobject streamApiInstance) {
     return (stream::StreamApiLow *) ctx.getObject(apiLong).getLongValue();
 }
 
-//extern "C"
-//JNIEXPORT jobject JNICALL
-//Java_com_simplito_java_privmx_1endpoint_modules_stream_StreamApiLow_init(
-//        JNIEnv *env,
-//        jobject thiz,
-//        jobject connection,
-//        jobject eventApi
-//) {
-//    JniContextUtils ctx(env);
-//    jobject result;
-//
-//    if (ctx.nullCheck(connection, "Connection") ||
-//            ctx.nullCheck(eventApi, "EventApi")) {
-//        return nullptr;
-//    }
-//
-////    ctx.callResultEndpointApi<jobject>(&result, [&ctx, &env, &thiz, &connection, &eventApi] {
-//        auto connection_c = getConnection(env, connection);
-//        auto eventApi_c = getEventApi(ctx, eventApi);
-//        auto streamApiLow = stream::StreamApiLow::create(
-//                *connection_c,
-//                *eventApi_c
-//                );
-//        auto streamApiLow_ptr = new stream::StreamApiLow();
-//        *streamApiLow_ptr = streamApiLow;
-//        return ctx.long2jLong((jlong) streamApiLow_ptr);
-////    });
-//    if (ctx->ExceptionCheck()) {
-//        return nullptr;
-//    }
-//    return result;
-//}
-
 extern "C"
 JNIEXPORT jobject JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_stream_StreamApiLow_create(
+Java_com_simplito_java_privmx_1endpoint_modules_stream_StreamApiLow_init(
         JNIEnv *env,
-        jclass clazz,
+        jobject thiz,
         jobject connection,
         jobject eventApi
 ) {
     JniContextUtils ctx(env);
-    if (ctx.nullCheck(connection, "Connection")) {
+    jobject result;
+
+    if (ctx.nullCheck(connection, "Connection") ||
+            ctx.nullCheck(eventApi, "EventApi")) {
         return nullptr;
     }
 
-//    jobject result;
-//    ctx.callResultEndpointApi<jobject>(&result, [&ctx, &env, &clazz, &connection, &eventApi] {
-    jmethodID initMID = ctx->GetMethodID(clazz, "<init>",
-            "(Ljava/lang/Long;)V");
-    auto connection_c = getConnection(env, connection);
-    auto eventApi_c = getEventApi(env, eventApi);
-    auto streamApiLow = StreamApiLow::create(
-            *connection_c,
-            *eventApi_c
-    );
-    auto *api = new StreamApiLow();
-    *api = streamApiLow;
-
-    jobject result = ctx->NewObject(
-            clazz,
-            initMID,
-            ctx.long2jLong((jlong) api)
-    );
-    return result;
-
-//    });
-    if (ctx->ExceptionCheck()) {
-        return nullptr;
-    }
-    return result;
+    ctx.callResultEndpointApi<jobject>(&result, [&ctx, &env, &connection, &eventApi] {
+        auto connection_c = getConnection(env, connection);
+        auto eventApi_c = getEventApi(env, eventApi);
+        auto streamApiLow = stream::StreamApiLow::create(
+                *connection_c,
+                *eventApi_c
+        );
+        auto streamApiLow_ptr = new stream::StreamApiLow();
+        *streamApiLow_ptr = streamApiLow;
+        return ctx.long2jLong((jlong) streamApiLow_ptr);
+    });
 }
 
 extern "C"
